Navbar rendering
In accordance with comments on PR #15, transitioned from a purely CSS based
method of displaying and hiding components to a method utilizing the
'useMediaQuery' hook provided by Material-UI.

This new implementation avoids rendering unnecessary components depending on the
screen width. Currently, this is done with a hardcoded value of 600, will later
have to make this the theme.breakpoints value when a themeprovider is
implemented.