StickyTimeSlot component
Sticky format of timeslots now placed into its own component to avoid conflicts
with HomePage's dependency on Timeslot.js.

May choose to merge these two components into one later on to avoid redundancy,
but for now this seems to be appropriate since the workshops list on the
HomePage differs from the list on the SchedulePage.